setting up workspace files should be the workspace's responsibility
