+++ src/workspace.rs
@@ -1,6 +1,14 @@
 use crate::job;
 use anyhow::{Context, Result};
+use path_absolutize::Absolutize;
+use std::fs;
 use std::path::{Path, PathBuf};
+
+#[cfg(target_family = "unix")]
+use std::os::unix::fs::symlink;
+
+#[cfg(target_family = "windows")]
+use std::os::windows::fs::symlink_file;
 
 #[derive(Debug)]
 pub struct Workspace(PathBuf);
@@ -12,6 +20,36 @@ impl Workspace {
         std::fs::create_dir_all(&workspace.0).context("could not create workspace")?;
 
         Ok(workspace)
+    }
+
+    pub fn set_up_files(&self, job: &job::Job) -> Result<()> {
+        for file in &job.input_files {
+            if let Some(parent_base) = file.parent() {
+                let parent = self.join(parent_base);
+
+                if !parent.exists() {
+                    fs::create_dir_all(parent).with_context(|| {
+                        format!("could not create parent for `{}`", file.display())
+                    })?;
+                }
+            }
+
+            let source = file.absolutize().with_context(|| {
+                format!("could not convert `{}` to an absolute path", file.display())
+            })?;
+
+            #[cfg(target_family = "unix")]
+            symlink(source, self.join(file)).with_context(|| {
+                format!("could not symlink `{}` into workspace", file.display())
+            })?;
+
+            #[cfg(target_family = "windows")]
+            symlink_file(source, workspace.join(file)).with_context(|| {
+                format!("could not symlink `{}` into workspace", file.display())
+            })?;
+        }
+
+        Ok(())
     }
 
     pub fn join<P: AsRef<Path>>(&self, other: P) -> PathBuf {
